Updated gateman controller

+++ app/Http/Controllers/GatemanController.php
@@ -49,7 +49,7 @@ class GatemanController extends Controller
 		else {
 	    	$users = [];
 
-	    	// get the resident's details and id requesting for the gateman
+	    	// get the resident's details requesting for the gateman together with the request id
 			foreach ($requests as $request) {
 		    	$user = User::join('resident_gateman', 'resident_gateman.user_id', '=', 'users.id')
 		    		->where('users.id', '=', $request->user_id)
@@ -84,8 +84,14 @@ class GatemanController extends Controller
         if (!$gateman) {
             return response()->json(['status' => false, 'message' => 'Request not found!'], 404);
         } else {
+	        // ensure that only the right gateman can accept the request
+	        if ($gateman->gateman_id != $gateman_id) {
+	            return response()->json(['status' => false, 'message' => 'Access denied'], 401);
+	        }
+
         	// check if the request has not been accepted
-        	$request = $gateman->where('gateman_id', $gateman_id)
+        	$request = $gateman->where('id', $id)
+        		->where('gateman_id', $gateman_id)
         		->where('request_status', 0)
         		->exists();
         	
@@ -96,7 +102,8 @@ class GatemanController extends Controller
         		if ($gateman->save()) {
 			        return response()->json([
 			        	'message' => 'The request has been accepted successfully',
-			        	'status' => true
+			        	'status' => true,
+			        	'resident_gateman' => $gateman
 			        ], 202);
         		} else {
 			        return response()->json([
@@ -128,15 +135,19 @@ class GatemanController extends Controller
         if (!$gateman) {
             return response()->json(['status' => false, 'message' => 'Request not found!'], 404);
         } else {
+	        // ensure that only the right gateman can reject the request
+	        if ($gateman->gateman_id != $gateman_id) {
+	            return response()->json(['status' => false, 'message' => 'Access denied'], 401);
+	        }
+
         	// check if the request has not been accepted
-        	$request = $gateman->where('gateman_id', $gateman_id)
+        	$request = $gateman->where('id', $id)
+        		->where('gateman_id', $gateman_id)
         		->where('request_status', 0)
         		->exists();
         	
-        	// update the request
         	if ($request) {
-        		$gateman->request_status = 1;
-
+	        	// reject the request
         		if ($gateman->destroy($id)) {
 			        return response()->json([
 			        	'message' => 'The request has been rejected successfully',
